pyproject.toml
[project]
name = "german-credit-risk-estimation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi-cache2>=0.2.2",
    "matplotlib>=3.10.6",
    "mlflow>=3.3.2",
    "notebook>=7.4.5",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pydantic>=2.11.7",
    "pytest>=8.4.2",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "yellowbrick>=1.5",
]
